+++ AutoSaver/_Thunderstore/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-Dec-09
+
+### Added
+
+- Support for saving extra Spike slots from [MoreBankTabs](https://thunderstore.io/c/atlyss/p/16MB/MoreBankTabs/) v0.1.0
+- Option to toggle autosaving whenever map instance changes.
+- Experimental feature that allows locally saving other player's saves in multiplayer.
+
+### Changed
+
+- Autosaves no longer happen whenever map instance changes, you must now toggle an option for this to happen.
+- Save backups now reference the slot that has been used. This allows multiple saves with the same character name to be backed up correctly.
+
 ## [1.0.0] - 2024-Dec-06
 
 ### Changed